Move source code files into src directory
