+++ .gitignore
@@ -1,3 +1,2 @@
-vendor
-.idea
-composer.lock
+/vendor/
+/composer.lock
