+++ src/Descriptor/TextDescriptor.php
